move ir_types_and_kwargs to common.py
